Add Archive Command
Previously, users could only delete tasks or mark them as done. Some users may want to archive tasks that could recur again. Tasks that get delayed could be kept for later as well to prevent cluttering of the task list.

To add this functionality, we add a new type of command object. We also add the corresponding messages. Finally, we allow our storage class to save the tasks to file accordingly. Methods in the TaskList and Storage class are also split to allow code sharing between different methods. This means we can add another useful delete all command using the new split classes.

This allows users to easily declutter their task list by archiving temporarily unused tasks. They could also easily bring back the archived tasks once they want to work on them again.

+++ src/main/java/me/yukun99/ip/commands/DeleteCommand.java
@@ -28,10 +28,17 @@ public class DeleteCommand extends Command {
     @Override
     public String getResponse() throws HelpBotInvalidTaskException, HelpBotIoException {
         List<Task> deletedTasks = new ArrayList<>();
-        if (args.length > 1) {
-            deletedTasks = taskList.deleteTasks(args);
-        } else {
-            deletedTasks.add(taskList.deleteTask(args[0]));
+        boolean deleteAll = false;
+        if (args[0].equals("all")) {
+            deletedTasks = taskList.deleteAll();
+            deleteAll = true;
+        }
+        if (!deleteAll) {
+            if (args.length > 1) {
+                deletedTasks = taskList.deleteTasks(args);
+            } else {
+                deletedTasks.add(taskList.deleteTask(args[0]));
+            }
         }
         storage.updateTasks();
         if (deletedTasks.size() > 1) {
